Revert "Revert "fix bugs""
This reverts commit dc9cb189ed97e2d5cfc09e78622d3c514fe4b9de.
