+++ src/pages/CheckCall.jsx
@@ -73,7 +73,7 @@ export default function CheckCall() {
             categories={el.categories}
             onDelete={() => {}}
             onDetail={() => {
-              navigate(`/check-call/${el.id}`);
+              navigate(`/check/${el.id}`);
             }}
           />
         ))}
